Add in templating for Jinja2
Updated employee_index.html to utilize Jinja templating to dynamically
build the Restrooms table viewed after logging in as an employee.

The employee_index.html file now iterates over query results returned
from Flask to build the table rows, as opposed to the static HTML
that was in use previously.