Add /employee_add route to app.py
Added the /employee_add route to app.py in order to handle both
the GET and POST HTTP methods for users navigating to /employee_add.

If a GET request, then simply return the page to the user where they
can fill out the form to add a new restroom.

If a POST request (via submit button being pressed on the form),
then collect the data provided and format into MySQL query to
insert new Restroom and associated data into DB.