Update employee_index and customer_index table layouts
After adding several rows to the employee_index table of Restrooms,
the table began bleeding over the edges and was not very aesthetic.

Added a div container around the <table> attribute with a static height
value and overflow attributes to handle scrolling. This seems to be
working much better.

Went ahead and did the same for customer_index.html