Add Functionality to Update Individual Rows from employee_index table
In order to POST to Flask, data needs to be in a form as opposed to
a table, which is currently how the data is laid out in our app.

Forms cannot be nested within a table. If we were to create one
form for all rows in the table, it'd be impossible to know which
row a user selected to update.

Thus, this commit adds functionality to employee_index.html to
dynamically create a quanitty of forms equal to the number of table
rows on the page. Each table cell within the individual rows then
references the appropriate form.

jQuery is then updated to specify the 'Update' button as
type="button" until the user clicks it. At that point, the button
value changes to "save" and upon selecting save, the button changes
to type="submit".

Data is POSTed to /employee_update, which currently just returns
the restroomID of the row that posted the update. The functionality
to update the full entry in MySQL as well as redirect the user back
to employee_index.html will be added in subsequent commits.