Implement DELETE Functionality
Edited employee_index.html to have the Delete button specify the
formaction attribute so that we can have two submit buttons that
belong to the same form while specifying different actions.

Created the /employee_delete route which deletes a row from the
employee_index.html page (default or filtered). This required an
update to the restroom_finder.sql scheme as the Reviews table was
missing a foreign key constraint to CASCASE on DELETE. That is to say,
when a Restroom is deleted, any reviews associated with the Restroom
are also deleted. Same for Locations and records in RestroomsEmployees.

The /employee_delete route redirects back to employee_index.html after
successful operation.